restructuring client/src directory
