+++ client/src/components/Cart/index.tsx
@@ -1,0 +1,2 @@
+import Cart from "./Cart";
+export default Cart;
