+++ client/src/components/ItemCard/index.tsx
@@ -1,0 +1,2 @@
+import ItemCard from "./ItemCard";
+export default ItemCard;
