+++ client/src/components/StoreHeader/index.tsx
@@ -1,0 +1,2 @@
+import StoreHeader from "./StoreHeader";
+export default StoreHeader;
