+++ client/src/components/StoreList/index.tsx
@@ -1,0 +1,2 @@
+import StoreList from "./StoreList";
+export default StoreList;
