+++ client/src/components/TextInputField/index.tsx
@@ -1,0 +1,2 @@
+import TextInputField from "./TextInputField";
+export default TextInputField;
