+++ client/src/interfaces.ts
