+++ client/src/pages/Home/index.tsx
@@ -1,0 +1,2 @@
+import Home from "./Home";
+export default Home;
