+++ client/src/pages/NotFound/index.tsx
@@ -1,0 +1,2 @@
+import NotFound from "./NotFound";
+export default NotFound;
