+++ client/src/pages/Receipt/index.tsx
@@ -1,0 +1,2 @@
+import Receipt from "./Receipt";
+export default Receipt;
