+++ client/src/pages/ScanStore/index.tsx
@@ -1,0 +1,2 @@
+import ScanStore from "./ScanStore";
+export default ScanStore;
